fix: components need createCompDef in header
