+++ sim/include/component/pos2DComponent.h
@@ -10,7 +10,21 @@ const char POS2D_COMPONENT_NAME[] = "Pos2DComponent";
 
 namespace {
 
-bento::protos::ComponentDef createCompDef();
+// This needs to be in the header as it is called immediately in the header as
+// well.
+bento::protos::ComponentDef createCompDef() {
+    auto compDef = bento::protos::ComponentDef();
+    compDef.set_name(ics::component::POS2D_COMPONENT_NAME);
+
+    auto& schema = *compDef.mutable_schema();
+    schema["x"] = bento::protos::Type();
+    schema["x"].set_primitive(bento::protos::Type_Primitive_INT64);
+
+    schema["y"] = bento::protos::Type();
+    schema["y"].set_primitive(bento::protos::Type_Primitive_INT64);
+
+    return compDef;
+};
 
 }  // namespace
