+++ sim/include/component/textureComponent.h
@@ -10,7 +10,18 @@ const char TEXTURE2D_COMPONENT_NAME[] = "Texture2DComponent";
 
 namespace {
 
-bento::protos::ComponentDef createCompDef();
+// This needs to be in the header as it is called immediately in the header as
+// well.
+bento::protos::ComponentDef createCompDef() {
+    auto compDef = bento::protos::ComponentDef();
+    compDef.set_name(ics::component::TEXTURE2D_COMPONENT_NAME);
+
+    auto& schema = *compDef.mutable_schema();
+    schema["texture"] = bento::protos::Type();
+    schema["texture"].set_primitive(bento::protos::Type_Primitive_INT64);
+
+    return compDef;
+};
 
 }  // namespace
